pyproject.toml
[project]
name = "billbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.39.4",
    "google-genai>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "stripe>=12.3.0",
]
